clippy

+++ ciphera/pkg/contracts/src/rollup.rs
@@ -296,7 +296,7 @@ impl RollupContract {
     pub async fn from_eth_node(eth_node: &EthNode, secret_key: SecretKey) -> Result<Self> {
         let rollup_addr = "cf7ed3acca5a467e9e704c703e8d87f634fb0fc9";
         let client = Client::from_eth_node(eth_node);
-        let chain_id = 5655 as u64;
+        let chain_id = 5655_u64;
         Self::load(client, &chain_id, rollup_addr, secret_key).await
     }
